requirements.txt
streamlit
pytesseract
pdf2image
openpyxl
pillow